+++ plugins/plugin-reputation/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @fayz-ai/plugin-reputation
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @fayz-ai/core@0.5.0
+  - @fayz-ai/ui@0.5.0
+
 ## 0.2.0
 
 ### Minor Changes